+++ my-app/my-app/package.json
@@ -14,10 +14,12 @@
     "graphql": "^16.11.0",
     "next": "14.2.32",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "swiper": "^12.0.1"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "6.0.0",
+    "@graphql-codegen/client-preset": "5.0.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
@@ -25,7 +27,6 @@
     "eslint-config-next": "14.2.32",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5",
-    "@graphql-codegen/client-preset": "5.0.0"
+    "typescript": "^5"
   }
 }
