+++ my-app/my-app/src/app/boards/page.tsx
@@ -1,10 +1,12 @@
 "use client";
 
-import BoardList from "../../components/boards-list/index";
+import BannerList from "@/components/boards-list/banner/index";
+import BoardList from "../../components/boards-list/list/index";
 
 export default function BoardsPage() {
   return (
     <>
+      <BannerList />
       <BoardList />
     </>
   );
