+++ my-app/my-app/src/components/boards-detail/comment-list/hook.ts
@@ -1,0 +1,17 @@
+"use client";
+
+import { FetchBoardCommentsDocument } from "@/commons/gql/graphql";
+import { useQuery } from "@apollo/client";
+import { useParams } from "next/navigation";
+
+export const useCommentList = () => {
+  const params = useParams();
+  const { data } = useQuery(FetchBoardCommentsDocument, {
+    variables: { boardId: String(params.boardId) },
+  });
+  console.log(data);
+
+  return {
+    data,
+  };
+};
