+++ my-app/my-app/src/components/boards-detail/comment-list/queries.ts
@@ -1,0 +1,15 @@
+import { gql } from "@apollo/client";
+
+export const FETCH_BOARD_COMMENTS = gql`
+  query FetchBoardComments($boardId: ID!) {
+    fetchBoardComments(boardId: $boardId) {
+      _id
+      writer
+      contents
+      rating
+      createdAt
+      updatedAt
+      deletedAt
+    }
+  }
+`;
