+++ my-app/my-app/src/components/boards-list/banner/styles.module.css
@@ -1,0 +1,6 @@
+.SwiperSlideImg {
+  max-width: 123rem;
+  width: 100%;
+  max-height: 36rem;
+  height: 100%;
+}
