+++ my-app/my-app/src/components/boards-list/list/styles.module.css
@@ -1,3 +1,16 @@
+.BoardContainer {
+  display: flex;
+  width: 100vw;
+  height: 100vh;
+  justify-content: center;
+}
+
+.BoardBody {
+  display: flex;
+  max-width: 120rem;
+  width: 100%;
+}
+
 .BoardFrame {
   display: flex;
   flex-direction: column;
@@ -5,13 +18,13 @@
   max-width: 80rem;
   min-width: 42.5rem;
   width: 100%;
+  height: 41rem;
   padding: 1.5rem 3rem;
   margin: 2.5rem 20rem;
   gap: 1.5rem;
   justify-content: center;
   align-content: center;
   /* align-self: stretch; */
-
   border-radius: 1rem;
   background: #fff;
   box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.08);
@@ -34,7 +47,7 @@
   gap: 0.625rem;
 }
 
-.BoardFrame_Header_title {
+.BoardFrame_Headxer_title {
   display: flex;
   max-width: 53rem;
   width: 100%;
@@ -74,20 +87,15 @@
   max-width: 74rem;
   width: 100%;
   height: 34.3rem;
-  /* 
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 1;
-  flex: 1 0 0; */
-  display: grid;
+  display: flex;
+  flex-direction: column;
 }
 
 .FetchBoard_list {
   display: flex;
-  width: 74rem;
-  /* width: 100%; */
+  width: 100%;
+  min-width: 74rem;
   height: 2.75rem;
-
   padding: 0.75rem 1.5rem;
   align-items: center;
   gap: 0.5rem;
@@ -135,8 +143,7 @@
   display: flex;
   max-width: 6.25rem;
   width: 100%;
-  justify-content: center;
-  align-items: center;
+  justify-content: left;
   gap: 0.625rem;
 
   overflow: hidden;
